+++ foxes/core/__init__.py
@@ -19,4 +19,3 @@ from .wake_frame import WakeFrame
 from .wake_model import WakeModel
 from .wake_superposition import WakeSuperposition
 from .vertical_profile import VerticalProfile
-from .runner import Runner
