--- foxes/models/runners/__init__.py
@@ -1,2 +1,0 @@
-from .simple_runner import SerialRunner
-from .dask_runner import DaskRunner
